Added RK3326_720 build target

+++ packages/amberelec/package.mk
@@ -67,7 +67,7 @@ makeinstall_target() {
     rm -rf ${INSTALL}/usr/config/distribution/modules/display_fix.sh
   fi
 
-  if [ ! "${DEVICE}" == "RG351MP" ]; then
+  if [ ! "${DEVICE}" == "RG351MP" ] && [ ! "${DEVICE}" == "RK3326_720" ]; then
     rm -rf ${INSTALL}/usr/config/distribution/modules/joyleds_conf.sh
   fi
 
@@ -76,7 +76,7 @@ makeinstall_target() {
 
   if [ "${DEVICE}" == "RG351P" ]; then
     cp ${INSTALL}/usr/config/distribution/configs/distribution.conf.351p ${INSTALL}/usr/config/distribution/configs/distribution.conf
-  elif  [ "${DEVICE}" == "RG351V" ] || [ "${DEVICE}" == "RG351MP" ]; then
+  elif  [ "${DEVICE}" == "RG351V" ] || [ "${DEVICE}" == "RG351MP" ] || [ "${DEVICE}" == "RK3326_720" ]; then
     cp ${INSTALL}/usr/config/distribution/configs/distribution.conf.351v ${INSTALL}/usr/config/distribution/configs/distribution.conf
   elif [ "${DEVICE}" == "RG552" ]; then
     cp ${INSTALL}/usr/config/distribution/configs/distribution.conf.552  ${INSTALL}/usr/config/distribution/configs/distribution.conf
@@ -105,6 +105,8 @@ makeinstall_target() {
     cp -r ${PKG_DIR}/overlay-p/* ${INSTALL}/usr/share/retroarch-overlays
   elif [ "${DEVICE}" == "RG351V" ] || [ "${DEVICE}" == "RG351MP" ]; then
     cp -r ${PKG_DIR}/overlay-v/* ${INSTALL}/usr/share/retroarch-overlays
+  elif [ "${DEVICE}" == "RK3326_720" ]; then
+    cp -r ${PKG_DIR}/overlay-720/* ${INSTALL}/usr/share/retroarch-overlays
   elif [ "${DEVICE}" == "RG552" ]; then
     cp -r ${PKG_DIR}/overlay-552/* ${INSTALL}/usr/share/retroarch-overlays
   fi
@@ -141,7 +143,7 @@ post_install() {
     enable_service fan_control.service
   fi
 
-  if [[ "${DEVICE}" =~ "RG351" ]]; then
+  if [[ "${DEVICE}" =~ "RG351" ]] || [[ "${DEVICE}" == "RK3326_720" ]]; then
     cp -f  ${PKG_DIR}/clocks/RK3326/clocklimits ${INSTALL}/etc
   elif [[ "${DEVICE}" == "RG552" ]]; then
     cp -f  ${PKG_DIR}/clocks/RK3399/clocklimits ${INSTALL}/etc
